v0.2a (March 2005)
Changed the location of the MBI and associated data from just after the
kernel to just below the 640k memory hole, because Xen assumes they will be
below 1MB and might overwrite them otherwise.

Added support for providing the command-line at boot time, and for
splitting the boot command-line between the kernel and the first module.

Removed pack-as-multiboot option, which was really a debugging feature.